+++ text_editor.py
@@ -117,6 +117,7 @@ class TextEditor(
         self._hovered_worktree_row = None
         self._worktree_drag_origin = None
         self._worktree_drag_target = None
+        self._worktree_ctrl_drag_active = False
         self._name_dialog = None
         self._confirm_dialog = None